refactor: remove and deprecate use-expo
