+++ example/src/molecules/battery/use-battery-level.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Caption } from 'react-native-paper';
-import { useBatteryLevel } from 'use-expo';
+import { useBatteryLevel } from '@use-expo/battery';
 import { Example, Information, Link, Measurement, Page } from '../../atoms';
 import { MoleculeProps } from '../../providers/molecule';
 import { docs } from '../../providers/urls';
