+++ example/src/molecules/battery/use-battery.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Button, Caption } from 'react-native-paper';
-import { useBattery } from 'use-expo';
+import { useBattery } from '@use-expo/battery';
 import { Example, Information, Link, Measurement, Page, Space } from '../../atoms';
 import { MoleculeProps } from '../../providers/molecule';
 import { docs } from '../../providers/urls';
