+++ example/src/molecules/brightness/use-system-brightness.tsx
@@ -2,7 +2,8 @@ import React, { useCallback } from 'react';
 import { Slider } from 'react-native';
 import { Caption } from 'react-native-paper';
 import { SYSTEM_BRIGHTNESS } from 'expo-permissions';
-import { usePermissions, useSystemBrightness } from 'use-expo';
+import { useSystemBrightness } from '@use-expo/brightness';
+import { usePermissions } from '@use-expo/permissions';
 import { debounce, round } from 'lodash';
 import { Example, Information, Link, Page, MissingPermissions } from '../../atoms';
 import { MoleculeProps } from '../../providers/molecule';
