+++ example/src/molecules/sensors/use-device-motion.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Caption } from 'react-native-paper';
-import { useDeviceMotion } from 'use-expo';
+import { useDeviceMotion } from '@use-expo/sensors';
 import { Example, Information, Link, Measurement, Page } from '../../atoms';
 import { MoleculeProps } from '../../providers/molecule';
 import { docs } from '../../providers/urls';
