--- packages/use-expo/CHANGELOG.md
@@ -1,177 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# 3.0.0 (2020-05-03)
-
-
-
-# 2.0.0 (2020-04-05)
-
-
-### Bug Fixes
-
-* **web-browser:** add tests, fix lint errors and finish docs ([#187](https://github.com/bycedric/use-expo/issues/187)) ([3ddd739](https://github.com/bycedric/use-expo/commit/3ddd73901a1cd2e33e0d7d6c067dd0d88b71f1cf))
-
-
-### Code Refactoring
-
-* upgrade all packages to expo sdk 37 ([#185](https://github.com/bycedric/use-expo/issues/185)) ([083eca2](https://github.com/bycedric/use-expo/commit/083eca28c2271f6581b051e38652e51b4da8bfc9)), closes [#186](https://github.com/bycedric/use-expo/issues/186)
-
-
-### BREAKING CHANGES
-
-* this also bumps all peer dependencies to the bundled native modules for SDK 37.
-
-* refactor(battery): upgrade to expo sdk 37
-
-* refactor(brightness): upgrade to expo sdk 37
-
-* refactor(font): upgrade to expo sdk 37
-
-* refactor(permissions): upgrade to expo sdk 37
-
-* refactor(sensors): upgrade to expo sdk 37
-
-* refactor: upgrade monorepo to expo sdk 37
-
-* fix: use compatible semver range for react types
-
-
-
-## 1.0.1 (2020-02-25)
-
-
-
-# 1.0.0 (2020-01-23)
-
-
-
-## 0.10.1 (2019-09-24)
-
-
-
-# 0.10.0 (2019-09-24)
-
-
-### Features
-
-* add hooks for new battery api ([#40](https://github.com/bycedric/use-expo/issues/40)) ([02042a2](https://github.com/bycedric/use-expo/commit/02042a2f729739f669fb6cb48f946049ac7573fd))
-
-
-
-# 0.9.0 (2019-07-31)
-
-
-
-# 0.8.0 (2019-07-07)
-
-
-### Features
-
-* add first draft of fonts hook ([#20](https://github.com/bycedric/use-expo/issues/20)) ([14c42fa](https://github.com/bycedric/use-expo/commit/14c42fa0d8e89bcd2676f6a9dcbec7e06d7aa32e))
-
-
-
-## 0.7.1 (2019-06-26)
-
-
-
-# 0.7.0 (2019-06-26)
-
-
-### Features
-
-* **example:** add example app to play with hooks ([#17](https://github.com/bycedric/use-expo/issues/17)) ([6bf4133](https://github.com/bycedric/use-expo/commit/6bf4133ece127167350d95a26f70a964a5163366))
-
-
-
-## 0.6.1 (2019-06-19)
-
-
-
-# 0.6.0 (2019-06-16)
-
-
-### Features
-
-* **screen-orientation:** add first draft of screen orientation hooks ([6376e1f](https://github.com/bycedric/use-expo/commit/6376e1fa4780c4df9f7fa7ffab098cfd70f24616))
-
-
-
-# 0.5.0 (2019-06-12)
-
-
-### Bug Fixes
-
-* add brightness to use expo package ([1920cad](https://github.com/bycedric/use-expo/commit/1920cadf96de5619de5b5805240d2fe01595bcb5))
-
-
-
-# 0.4.0 (2019-06-11)
-
-
-### Features
-
-* add eslint to all packages ([#8](https://github.com/bycedric/use-expo/issues/8)) ([7008607](https://github.com/bycedric/use-expo/commit/7008607d0f63df5baf315f8aba29fbdbd888da45))
-
-
-
-# 0.3.0 (2019-06-10)
-
-
-### Features
-
-* add use expo alias library to install all ([#6](https://github.com/bycedric/use-expo/issues/6)) ([98bcdf7](https://github.com/bycedric/use-expo/commit/98bcdf72409ceb820ffda13f16d7189d8bace1af))
-
-
-
-
-
-# [2.0.0](https://github.com/bycedric/use-expo/compare/1.0.1...2.0.0) (2020-04-05)
-
-
-### Bug Fixes
-
-* **web-browser:** add tests, fix lint errors and finish docs ([#187](https://github.com/bycedric/use-expo/issues/187)) ([3ddd739](https://github.com/bycedric/use-expo/commit/3ddd73901a1cd2e33e0d7d6c067dd0d88b71f1cf))
-
-
-### Code Refactoring
-
-* upgrade all packages to expo sdk 37 ([#185](https://github.com/bycedric/use-expo/issues/185)) ([083eca2](https://github.com/bycedric/use-expo/commit/083eca28c2271f6581b051e38652e51b4da8bfc9)), closes [#186](https://github.com/bycedric/use-expo/issues/186)
-
-
-### BREAKING CHANGES
-
-* this also bumps all peer dependencies to the bundled native modules for SDK 37.
-
-* refactor(battery): upgrade to expo sdk 37
-
-* refactor(brightness): upgrade to expo sdk 37
-
-* refactor(font): upgrade to expo sdk 37
-
-* refactor(permissions): upgrade to expo sdk 37
-
-* refactor(sensors): upgrade to expo sdk 37
-
-* refactor: upgrade monorepo to expo sdk 37
-
-* fix: use compatible semver range for react types
-
-
-
-
-
-## [1.0.1](https://github.com/bycedric/use-expo/compare/1.0.0...1.0.1) (2020-02-25)
-
-**Note:** Version bump only for package use-expo
-
-
-
-
-
-# [1.0.0](https://github.com/bycedric/use-expo/compare/v0.10.1...1.0.0) (2020-01-23)
-
-**Note:** Version bump only for package use-expo
