--- packages/use-expo/README.md
@@ -1,32 +1,0 @@
-<div align="center">
-    <h1>expo hooks</h1>
-     <p>Complementary hooks for <a href="https://github.com/expo/expo">Expo</a></p>
-    <sup>
-        <a href="https://github.com/byCedric/use-expo/blob/master/packages/use-expo/CHANGELOG.md">
-            <img src="https://img.shields.io/npm/v/@use-expo/use-expo?style=flat-square" alt="changelog" />
-        </a>
-        <a href="https://github.com/bycedric/use-expo/actions">
-            <img src="https://img.shields.io/github/workflow/status/byCedric/use-expo/Packages/master.svg?style=flat-square" alt="builds" />
-        </a>
-        <a href="https://exp.host/@bycedric/use-expo">
-            <img src="https://img.shields.io/badge/demo-expo.io-lightgrey.svg?style=flat-square" alt="demo" />
-        </a>
-    </sup>
-    <br />
-    <p align="center">
-        <a href="https://github.com/byCedric/use-expo#readme"><b>All hooks</b></a>
-        &nbsp;&nbsp;&mdash;&nbsp;&nbsp;
-        <a href="https://github.com/byCedric/use-expo#usage"><b>Usage</b></a>
-        &nbsp;&nbsp;&mdash;&nbsp;&nbsp;
-        <a href="https://github.com/byCedric/use-expo/blob/master/CHANGELOG.md"><b>Changelog</b></a>
-    </p>
-    <br />
-    <pre>expo install use-expo</pre>
-    <br />
-</div>
-
-<div align="center">
-    <br />
-    with :heart: <strong>byCedric</strong>
-    <br />
-</div>
