--- packages/use-expo/package.json
@@ -1,62 +1,0 @@
-{
-	"name": "use-expo",
-	"version": "3.0.0",
-	"description": "Complementary hooks for expo",
-	"keywords": [
-		"expo",
-		"hooks"
-	],
-	"license": "MIT",
-	"author": "Cedric van Putten <me@bycedric.com> (https://bycedric.com)",
-	"homepage": "https://github.com/bycedric/use-expo/tree/master/packages/use-expo#readme",
-	"bugs": {
-		"url": "https://github.com/bycedric/use-expo/issues"
-	},
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/bycedric/use-expo.git",
-		"directory": "packages/use-expo"
-	},
-	"directories": {
-		"lib": "src",
-		"test": "tests"
-	},
-	"main": "build/commonjs/index.js",
-	"react-native": "src/index.ts",
-	"module": "build/module/index.js",
-	"types": "build/typescript/src/index.d.ts",
-	"files": [
-		"build",
-		"src"
-	],
-	"scripts": {
-		"prepare": "bob build",
-		"build": "bob build"
-	},
-	"dependencies": {
-		"@use-expo/application": "^1.0.0",
-		"@use-expo/battery": "^2.0.0",
-		"@use-expo/brightness": "^2.0.0",
-		"@use-expo/font": "^2.0.0",
-		"@use-expo/permissions": "^2.0.0",
-		"@use-expo/screen-orientation": "^2.0.0",
-		"@use-expo/sensors": "^2.0.0",
-		"@use-expo/web-browser": "^2.0.0"
-	},
-	"devDependencies": {
-		"@react-native-community/bob": "^0.10.0",
-		"typescript": "^3.8.3"
-	},
-	"@react-native-community/bob": {
-		"source": "src",
-		"output": "build",
-		"targets": [
-			"commonjs",
-			"module",
-			"typescript"
-		]
-	},
-	"publishConfig": {
-		"access": "public"
-	}
-}
