--- packages/use-expo/src/index.ts
@@ -1,8 +1,0 @@
-export * from '@use-expo/application';
-export * from '@use-expo/battery';
-export * from '@use-expo/brightness';
-export * from '@use-expo/font';
-export * from '@use-expo/permissions';
-export * from '@use-expo/screen-orientation';
-export * from '@use-expo/sensors';
-export * from '@use-expo/web-browser';
