--- packages/use-expo/tsconfig.json
@@ -1,19 +1,0 @@
-{
-	"extends": "../../tsconfig.json",
-	"compilerOptions": {
-		"outDir": "./build/typescript"
-	},
-	"include": [
-		"src"
-	],
-	"references": [
-		{ "path": "../application" },
-		{ "path": "../battery" },
-		{ "path": "../brightness" },
-		{ "path": "../font" },
-		{ "path": "../permissions" },
-		{ "path": "../screen-orientation" },
-		{ "path": "../sensors" },
-		{ "path": "../web-browser" },
-	]
-}
